Removed some unnecessary code
I removed all the data arguments that were in the functions, they are not needed. I also removed the if statements in the while loop for .stop, .source and .update. They are also not needed since the while loop looks through the functions dictionary. Also fixed a typo in the functions dict.